code/requirements.txt
tensorflow-gpu==2.1.0
numpy==1.19.5
opencv-python==4.5.3.56
PyYAML==5.4.1
tqdm==4.62.2
Pillow==8.3.2
Cython==0.29.24
ipython==7.27.0
dlib==19.22.1